Multiple fixes on Xe::IO::IMemoryStream for invalid behaviors or crashes

+++ include/XeSDK/XeExtensionsRapidjsonReadStream.h
@@ -44,7 +44,7 @@ namespace rapidjson
 
 		size_t Tell() const
 		{
-			return m_Stream->GetPosition();
+			return (size_t)m_Stream->GetPosition();
 		}
 
 		// Not implemented
